pyproject.toml
[tool.poetry]
name = "analytics"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.8"
dbt-bigquery = "^1.3.0"
flask = "^2.2.2"
gunicorn = "^20.1.0"
google-cloud-logging = "^3.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
